pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-application-git-tagging"
version = "0.0.1a3"
description = "Application layer for pythoneda-artifact/git-tagging"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-application/git-tagging"
packages = [{include = "pythonedaartifactapplicationgittagging"}]

[tool.poetry.dependencies]
python = "^3.8"
pythoneda-base = "^0.0.1a15"
pythoneda-artifact-git-tagging = "^0.0.1a3"
pythoneda-artifact-infrastructure-git-tagging = "^0.0.1a3"
pythoneda-application-base = "^0.0.1a11"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
